+++ CHANGELOG.md
@@ -24,6 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#1242](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1242))
 - `opentelemetry-util-http` Add support for sanitizing HTTP header values.
   ([#1253](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1253))
+- Add metric instrumentation in starlette
+  ([#1327](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1327))
+  
 
 ### Fixed